Move module to be a standalone plugin
