+++ build.gradle
@@ -1,17 +1,16 @@
-import dev.tigrao.bountyhunter.plugin.TrackerPlugin
-
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
-apply plugin: TrackerPlugin
-
 buildscript {
+
     ext.kotlin_version = '1.3.61'
     repositories {
+        mavenLocal()
         google()
         jcenter()
-
     }
+
     dependencies {
+        classpath 'dev.tigrao:tracker:0.0.1'
         classpath 'com.android.tools.build:gradle:3.5.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // NOTE: Do not place your application dependencies here; they belong
@@ -30,3 +29,5 @@ allprojects {
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
+
+apply plugin: 'BountyHunter'
