+++ settings.gradle
@@ -2,4 +2,6 @@ include ':app'
 include ':network'
 include ':core'
 include ':sample:core'
+
+include ':tracker'
 rootProject.name = 'BountyHunter'
