+++ tracker/src/main/java/dev/tigrao/bountyhunter/tracker/TrackerTask.kt
@@ -1,4 +1,4 @@
-package dev.tigrao.bountyhunter.plugin
+package dev.tigrao.bountyhunter.tracker
 
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
