+++ tracker/src/main/resources/META-INF/gradle-plugins/BountyHunter.properties
@@ -1,0 +1,1 @@
+implementation-class=dev.tigrao.bountyhunter.tracker.TrackerPlugin
